wrote new database for storing the library.  we now use a hash table based on word tokens in the filename, with a list of files for each word
modified malloc debugger to use a hash table to store the allocation
information because a single list was too slow for large amounts of data

modified buffer_queue() to allocate chunks of memory and allocate a new
buffer in the list if there is not enough room to append, instead of using
realloc.  this should improve performace

began adding extra checks for when we run out of memory so that we can
recover gracefully.