added timestamps to the channel modes on synch
changed the .created member of USERDB to be .timestamp.  This field is now
used to keep track of when the entry was last modified so it can be used to
synch servers better.

fixed synch of cloak status.  added ability to absolutely turn cloak on in a
server->server message

rewrote level() to be much cleaner

muzzle status was not correctly restored from the users file upon startup

new function create_db() created a USERDB struct based on a USER struct to
avoid duplicate code in level() and muzzle()