check for -lz in /usr/local/lib
cast pid_t to int in dump_pid()

added eject_client() to find a client with 0 files shared that has been
connected the longest and kill it.  if the server is full, login() will call
eject_client() to try and kill someone else to make room for the current
login.